requirements.txt
dataclasses_json
ijson
pathvalidate
pyyaml
colorlog